// src/enums/origem-da-mercadoria.enum.ts
export enum OrigemMercadoriaEnum {
  NACIONAL = 0,
  ESTRANGEIRA_IMPORTACAO_DIRETA = 1,
  ESTRANGEIRA_ADQUIRIDA_MERCADO_INTERNO = 2,
  NACIONAL_MERCADORIA_IMPORTACAO_ENTRE_40_E_70_PORCENTO = 3,
  NACIONAL_PRODUCAO_PROCESSOS_PRODUTIVOS_BASICOS = 4,
  NACIONAL_MERCADORIA_IMPORTACAO_ATE_40_PORCENTO = 5,
  ESTRANGEIRA_IMPORTACAO_DIRETA_SEM_SIMILAR_NACIONAL = 6,
  ESTRANGEIRA_ADQUIRIDA_MERCADO_INTERNO_SEM_SIMILAR_NACIONAL = 7,
  NACIONAL_MERCADORIA_IMPORTACAO_ACIMA_70_PORCENTO = 8,
}